upscalers/requirements.txt
pika
python-dotenv
Pillow
jsonschema
sdkit